+++ ranvier.json
@@ -14,12 +14,10 @@
     "ranvier-commands",
     "ranvier-effects",
     "ranvier-groups",
-    //"ranvier-input-events",
     "myelin-input-events",
     "ranvier-player-events",
     "ranvier-skills",
     "ranvier-combat",
-    //"ranvier-classes",
     "myelin-classes",
     "ranvier-crafting",
     "ranvier-quests",
